extract useIsMounted
