+++ packages/react-google-maps-api/src/utils/useIsMounted.tsx
@@ -1,0 +1,14 @@
+import * as React from 'react'
+
+const useIsMounted = () => {
+  const isMountedRef = React.useRef(false)
+  React.useEffect(function trackMountedState() {
+    isMountedRef.current = true
+    return () => {
+      isMountedRef.current = false
+    }
+  }, [])
+  return () => isMountedRef.current
+}
+
+export default useIsMounted;
